separate recipe into different components
